+++ vop.py
@@ -288,6 +288,18 @@ def measure_noise():
     # Returns a 200 OK to the Javascript fetch so the UI knows the command was accepted
     return jsonify({"status": "started"})
 
+@app.route('/map_hot_pixels', methods=['POST'])
+def map_hot_pixels():
+    dispatch_engine('map_hot_pixels', request.json)
+    return jsonify({"status": "started"})
+
+@app.route('/nuke_hot_pixels', methods="['POST']")
+def nuke_hot_pixels():
+    hp_file = os.path.join(BASE_DIR, "static", "hot_pixels.json")
+    if os.path.exists(hp_file):
+        os.remove(hp_file)
+    return jsonify({"status": "clear"})
+
 @app.route('/panic', methods=['POST'])
 def panic():
     print("[VOP SERVER] ACTION: PANIC STOP")